Vc to submit SyncCommitteeSignature on early block (#4615)
Vc to submit SyncCommitteeSignature asap
