+++ packages/validator/src/validator.ts
@@ -125,6 +125,7 @@ export class Validator {
       api,
       clock,
       validatorStore,
+      emitter,
       chainHeaderTracker,
       metrics
     );
